move TimeseriesMetadataService under service folder
