+++ bundles/framework/timeseries/service/TimeseriesMetadataService.js
